backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.1
Pillow==10.2.0
python-dotenv==1.0.1
drf-yasg==1.21.7
